fix: isolate BaseFactory class
`_BaseFactory` is to be shared with other modules, so it deserves its own home.
