+++ BFAIR/io/_path.py
@@ -17,7 +17,6 @@ def static_path(*args) -> str:
     Returns
     -------
     str
-        Path to the static folder.
     """
     # Output must be str to be compatible with cobra/pytfa
     return str(Path(static.__file__).parent.joinpath(*args))
